+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 
 Added
 -----
+- Add support for ``Python 3.11``
 
 Changed
 -------